Cleanup: restructure project for better organization and use dotenv when testing
