+++ .env-example
@@ -1,0 +1,2 @@
+OPENSKY_USER=your-username
+OPENSKY_PASS=your-password
