+++ .gitignore
@@ -15,3 +15,6 @@ Cargo.lock
 /target
 
 *.vim
+
+# Ignore the .env file
+.env
