+++ Cargo.toml
@@ -12,6 +12,7 @@ reqwest = "0.11"
 thiserror = "1.0"
 serde = { version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.68"
+dotenv = "0.15.0"
 
 [dev-dependencies]
 tokio = { version = "1.12", features = ["full"] }
